NOJIRA enforce scala 3 syntax (#102)
While making another PR, I used an underscore wildcard import before
realising the repo is on scala 3. I was surprised the CI was happy with
it.

This PR enables the rewrite flag, which means scalafmt will convert to
new syntax whenever it runs, including the wildcard imports. This is
customisable at a lower level, so we can disable individual conversions
if we like. I have left it on the default for now.